// backend/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  VICTIM: 'VICTIM',
  OFFICER: 'OFFICER',
  ADMIN: 'ADMIN'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const Gender = {
  MALE: 'MALE',
  FEMALE: 'FEMALE',
  OTHER: 'OTHER',
  PREFER_NOT_TO_SAY: 'PREFER_NOT_TO_SAY'
} as const

export type Gender = (typeof Gender)[keyof typeof Gender]


export const FIRStatus = {
  DRAFT: 'DRAFT',
  SUBMITTED: 'SUBMITTED',
  ACKNOWLEDGED: 'ACKNOWLEDGED',
  UNDER_INVESTIGATION: 'UNDER_INVESTIGATION',
  CHARGESHEET_FILED: 'CHARGESHEET_FILED',
  CLOSED: 'CLOSED',
  REJECTED: 'REJECTED'
} as const

export type FIRStatus = (typeof FIRStatus)[keyof typeof FIRStatus]


export const UrgencyLevel = {
  LOW: 'LOW',
  MEDIUM: 'MEDIUM',
  HIGH: 'HIGH',
  CRITICAL: 'CRITICAL'
} as const

export type UrgencyLevel = (typeof UrgencyLevel)[keyof typeof UrgencyLevel]


export const Language = {
  ENGLISH: 'ENGLISH',
  HINDI: 'HINDI',
  BHOJPURI: 'BHOJPURI',
  MARATHI: 'MARATHI',
  TAMIL: 'TAMIL',
  TELUGU: 'TELUGU',
  BENGALI: 'BENGALI',
  GUJARATI: 'GUJARATI',
  KANNADA: 'KANNADA',
  MALAYALAM: 'MALAYALAM',
  PUNJABI: 'PUNJABI',
  ODIA: 'ODIA'
} as const

export type Language = (typeof Language)[keyof typeof Language]


export const NotificationStatus = {
  PENDING: 'PENDING',
  SENT: 'SENT',
  FAILED: 'FAILED'
} as const

export type NotificationStatus = (typeof NotificationStatus)[keyof typeof NotificationStatus]


export const OfficerVerificationStatus = {
  PENDING: 'PENDING',
  VERIFIED: 'VERIFIED',
  REJECTED: 'REJECTED'
} as const

export type OfficerVerificationStatus = (typeof OfficerVerificationStatus)[keyof typeof OfficerVerificationStatus]


export const CrimeCategory = {
  VIOLENT: 'VIOLENT',
  PROPERTY: 'PROPERTY',
  CYBERCRIME: 'CYBERCRIME',
  DOMESTIC_VIOLENCE: 'DOMESTIC_VIOLENCE',
  FINANCIAL_FRAUD: 'FINANCIAL_FRAUD',
  SEXUAL_OFFENSE: 'SEXUAL_OFFENSE',
  PUBLIC_ORDER: 'PUBLIC_ORDER',
  DRUG_OFFENSE: 'DRUG_OFFENSE',
  OTHER: 'OTHER'
} as const

export type CrimeCategory = (typeof CrimeCategory)[keyof typeof CrimeCategory]